+++ commons-lang/src/main/java/com/hb0730/commons/lang/ClassUtils.java
@@ -2,6 +2,9 @@ package com.hb0730.commons.lang;
 
 import com.hb0730.commons.lang.collection.ArrayUtils;
 import com.hb0730.commons.lang.convert.BasicTypeEnum;
+import com.hb0730.commons.lang.reflect.TypeUtils;
+
+import java.lang.reflect.Type;
 
 /**
  * 类工具类
@@ -25,6 +28,45 @@ public class ClassUtils {
         } catch (Throwable ex) {
             return false;
         }
+    }
+
+
+    /**
+     * 判断类是否为枚举类型
+     *
+     * @param clazz 类
+     * @return 是否为枚举类型
+     * @since 1.0.2
+     */
+    public static boolean isEnum(Class<?> clazz) {
+        return null != clazz && clazz.isEnum();
+    }
+
+    /**
+     * 获得给定类的第一个泛型参数
+     *
+     * @param clazz 被检查的类，必须是已经确定泛型类型的类
+     * @return {@link Class}
+     * @since 1.0.2
+     */
+    public static Class<?> getTypeArgument(Class<?> clazz) {
+        return getTypeArgument(clazz, 0);
+    }
+
+    /**
+     * 获得给定类的泛型参数
+     *
+     * @param clazz 被检查的类，必须是已经确定泛型类型的类
+     * @param index 泛型类型的索引号，即第几个泛型类型
+     * @return {@link Class}
+     * @since 1.0.2
+     */
+    public static Class<?> getTypeArgument(Class<?> clazz, int index) {
+        final Type argumentType = TypeUtils.getTypeArgument(clazz, index);
+        if (argumentType instanceof Class) {
+            return (Class<?>) argumentType;
+        }
+        return null;
     }
 
     /**