+++ CHANGELOG.md
@@ -3,6 +3,9 @@
  + add BasicTypeEnum 基本类型与包装类型 枚举
  + add ClassUtils#isAllAssignableFrom,ClassUtils#isBasicType等方法
  + add ReflectUtils#getMethod等方法
+ + add Converter 类型转换(bate)
+ + add classUtils#getTypeArgument,classUtils#isEnum等方法
+ + add TypeUtils类 和 ReflectUtils#getConstructor等方法
 ## 1.0.1
  ## fix
  + fix UrlUtils 参数判断